feat: improve compatibility with observable
In a part of working on #4165, there's several redirects and general
incompatibilities that Sponge can introduce with other mods. Notably
out of the box, there's Observable, which aims to provide per-chunk
based observability of things that are occurring. It is not confirmed
this alone is enough to make it work.